Test Combat Code
This is mainly a proof of concept code. I'm using it to test ideas and to get a feel for how the combat will play out. IT is very basic. The only attack option is 'punch'. I need to figure out how to get variables to be accessed by other scripts, for example, i need a script for when the character heals, which has to add to the playerHealth value in both all combat scripts. Also other attacks the player can make will need to use the actionPoints and the enemyHealth variables. Any help or suggestions is greatly appreciated.